+++ src/protocols/writing.rs
@@ -61,6 +61,10 @@ where
                                     Err(e) => {
                                         node.known_peers().register_failure(addr);
                                         error!(parent: node.span(), "couldn't send a message to {}: {}", addr, e);
+                                        if node.config().fatal_io_errors.contains(&e.kind()) {
+                                            node.disconnect(addr);
+                                            break;
+                                        }
                                     }
                                 }
                             } else {